+++ pages/_app.tsx
@@ -14,7 +14,7 @@ import { TransactionStore } from '@context/TransactionContext';
 import { FactoryStore } from '@context/FactoryContext';
 import GlobalStyles from 'styles/GlobalStyles';
 import styled from 'styled-components';
-import WhitelistBlock from '@components/WhitelistBlock';
+// import WhitelistBlock from '@components/WhitelistBlock';
 
 const USERSNAP_GLOBAL_API_KEY = process.env.NEXT_PUBLIC_USERSNAP_GLOBAL_API_KEY;
 const USERSNAP_API_KEY = process.env.NEXT_PUBLIC_USERSNAP_API_KEY;
@@ -84,13 +84,14 @@ const App = ({ Component, pageProps }: AppProps) => { // eslint-disable-line
                         <GraphProvider>
                             <FactoryStore>
                                 <TransactionStore>
-                                    {process.env.NEXT_PUBLIC_DEPLOYMENT === 'DEVELOPMENT' ? (
-                                        <Component {...pageProps} />
-                                    ) : (
-                                        // <WhitelistBlock>
-                                            <Component {...pageProps} />
-                                        // </WhitelistBlock>
-                                    )}
+                                    <Component {...pageProps} />
+                                    {/*{process.env.NEXT_PUBLIC_DEPLOYMENT === 'DEVELOPMENT' ? (*/}
+                                    {/*    <Component {...pageProps} />*/}
+                                    {/*) : (*/}
+                                    {/*    <WhitelistBlock>*/}
+                                    {/*        <Component {...pageProps} />*/}
+                                    {/*    </WhitelistBlock>*/}
+                                    {/*)}*/}
                                 </TransactionStore>
                             </FactoryStore>
                         </GraphProvider>
